refactor: modify types route
