+++ src/types/score/index.ts
